+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 * Issue #43 https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/43
 * Added Terminology API and mockked for tests
+* Updated toggle buttons https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/38